Reorganize marmara→bossphorus
