--- bossphorus.code-workspace
@@ -1,14 +1,0 @@
-{
-	"folders": [
-		{
-			"path": ".circleci"
-		},
-		{
-			"path": "bossphorus"
-		},
-		{
-			"path": "marmara"
-		}
-	],
-	"settings": {}
-}
