+++ bossphorus/marmara/marmara.py
@@ -1,0 +1,16 @@
+import numpy as np
+
+from .engine import StorageEngine
+
+class Marmara:
+    def __init__(self, storage_engine: StorageEngine) -> None:
+        self.storage_engine = storage_engine
+
+    def get():
+        ...
+
+    def has():
+        ...
+
+    def put():
+        ...
