--- marmara/marmara/__init__.py
@@ -1,6 +1,0 @@
-from . import coordinate_frame
-from . import engine
-from .marmara import Marmara
-from . import version
-
-__version__ = version.__version__
