--- marmara/marmara/marmara.py
@@ -1,7 +1,0 @@
-
-import numpy as np
-import zmq
-
-class Marmara:
-    def __init__(self) -> None:
-        pass
