--- marmara/marmara/version.py
@@ -1,3 +1,0 @@
-"""Declare package version."""
-
-__version__ = "1.0.0"
